Pieced together nice model separation

+++ Kosystem.Repository.EF/DbPerson.cs
@@ -14,6 +14,6 @@ namespace Kosystem.Repository.EF
         public DateTime? EnqueuedAt { get; set; }
 
         public DbRoom? Room { get; set; }
-        public int? RoomId { get; set; }
+        public long? RoomId { get; set; }
     }
 }
