+++ Kosystem.Shared/UpdatePersonModel.cs
@@ -1,6 +1,6 @@
 namespace Kosystem.Shared
 {
-    public record UpdatePersonModel(int Id, string Name)
+    public record UpdatePersonModel(long Id, string Name)
     {
         // No additional properties so far.
     }
